Rename plugin to rre-maven-generic-search-plugin.
