+++ rre-maven-plugin/rre-maven-generic-search-plugin/pom.xml
@@ -26,7 +26,7 @@
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>rre-maven-evaluation-plugin</artifactId>
+    <artifactId>rre-maven-generic-search-plugin</artifactId>
     <packaging>maven-plugin</packaging>
     <version>1.0</version>
     <name>RRE - Maven Evaluation Plugin</name>
